added channel logic and restructured app for easier testing
